refactor: useRegister 페이지별 분리 및 event 전달방식 개선
