+++ client/src/components/Register/RegisterAuthForm/RegisterAuthForm.tsx
@@ -1,8 +1,6 @@
 import { VFC } from 'react';
-import { useHistory } from 'react-router-dom';
 
 import Anchor from '../../@atom/Anchor/Anchor';
-
 import {
   RegisterAnchorContainer,
   RegisterButtonContainer,
@@ -12,35 +10,10 @@ import {
   NaverButton,
 } from './RegisterAuthForm.styles';
 import { OAUTH } from '../../../constants/constant';
-
-enum OAuthSites {
-  google,
-  naver,
-  kakao,
-}
+import useRegisterAuth from '../../../service/hooks/useRegisterAuth';
 
 const RegisterAuthForm: VFC = () => {
-  const history = useHistory();
-
-  const openOAuthPage = (name: keyof typeof OAuthSites) => {
-    switch (name) {
-      case OAUTH.GOOGLE.NAME:
-        window.open(OAUTH.GOOGLE.URL, '_target');
-        break;
-
-      case OAUTH.NAVER.NAME:
-        // (다음 페이지로 넘어가기 위한 임시 코드로 대체)
-        // window.open(OAUTH.NAVER, '_target');
-        history.push('/register/url');
-        break;
-
-      case OAUTH.KAKAO.NAME:
-        window.open(OAUTH.KAKAO.URL, '_target');
-        break;
-
-      default:
-    }
-  };
+  const { openOAuthPage } = useRegisterAuth();
 
   return (
     <>
